Adding sudoku 2, removing minisat
TODO: re-implement 3x3 box rules in sudoku2

I've removed minisat from the repo since it is available online and in binary
form elsewhere for other platforms.

I figured out our "clause" issue. We were breaking the cnf file down into a
bunch of little chunks which had 'p cnf {} {}', but the sat solver wants
everything as one. We wanted to take all of the clauses and add them together
and pass everything at once to 'p cnf {} {}' instead of having many.

I've added an argument parser to the sudoku2

./sudoku2 <inputfile> <outputfile> <minisatexecpath> <tmp_in> <tmp_out>

I'm considering having the program just create the temp files in /tmp (though
this is not cross-compatible for windows machines) and only having the first
three arguments.