Added more test sections
We don't have to do all of those tests, but I think we should do at least 3 or 4.
I also added a discussion and some generic statements in the discussion and the conclusion to be completed after the tests (which, based on our methods, have to be done on Evan's computer). If the conclusion was supposed to contain the discussion points, then I apologize and will move the discussion stuff to the conclusion.
Also, I made test 1 about medium puzzles instead of random puzzles which makes more sense, I think.